refactored paths
