+++ src/diff/mod.rs
@@ -1,0 +1,7 @@
+pub mod diff;
+pub mod set_diff;
+pub mod vec_diff;
+
+pub use diff::Diff;
+pub use set_diff::SetDiff;
+pub use vec_diff::VecDiff;
