+++ src/handle/mod.rs
@@ -1,0 +1,9 @@
+pub mod tag;
+pub mod address;
+pub mod identity;
+pub mod location;
+
+pub use tag::Tag;
+pub use address::Address;
+pub use identity::Identity;
+pub use location::Location;
